[YS-85] refactor: popover 에서 filter 로 네이밍 수정
