+++ src/app/home/components/PostContainer/FilterContainer/FilterContainer.tsx
@@ -1,14 +1,14 @@
+import AreaFilter from '../AreaFilter/AreaFilter';
+import ContactTargetFilter from '../ContactTargetPopover/ContactTargetFilter';
+import ProgressMethodFilter from '../ProgressMethodFilter/ProgressMethodFilter';
 import { filterLayout } from './FilterContainer.styles';
-import ContactTargetPopover from '../ContactTargetPopover/ContactTargetPopover';
-import ProgressMethod from '../ProgressMethod/ProgressMethod';
-import RegionPopover from '../RegionPopover/RegionPopover';
 
 const FilterContainer = () => {
   return (
     <div css={filterLayout}>
-      <ProgressMethod />
-      <ContactTargetPopover />
-      <RegionPopover />
+      <ProgressMethodFilter />
+      <ContactTargetFilter />
+      <AreaFilter />
     </div>
   );
 };
