+++ src/app/home/components/PostContainer/ProgressMethodFilter/ProgressMethodFilter.styles.ts
